refactor: separate AdaptiveJourneyScreen from quiz_screen.dart
Moved AdaptiveJourneyScreen to dedicated feature directory to improve maintainability and follow Single Responsibility Principle.

Changes:
- Created lib/features/adaptive_journey/ directory structure
- Moved AdaptiveJourneyScreen + _AdaptiveJourneyScreenState
- Extracted ModuleTileState to separate model file
- Extracted LessonCard to separate widget file
- Updated router.dart imports
- quiz_screen.dart reduced from 2000+ to ~600 lines

Architecture improvements:
- Clear separation of concerns
- Easier to maintain and test
- Follows Flutter best practices
- Prep work for future Riverpod migration

Verification:
- flutter analyze:  No issues
- Build:  Success
- Routing:  Working
